requirements.txt
anyio==2.0.2
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.2.1
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cycler==0.10.0
cymem==2.0.5
decorator==4.4.2
defusedxml==0.6.0
descartes==1.1.0
dill==0.3.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
funcy==1.15
future==0.18.2
gensim==3.8.3
idna==2.10
importlib-metadata==3.3.0
iniconfig==1.1.1
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyter-server==1.2.0
jupyterlab==2.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mizani==0.7.2
murmurhash==1.0.5
nbclassic==0.2.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
nltk==3.5
notebook==6.1.6
numexpr==2.7.2
numpy==1.19.5
packaging==20.8
palettable==3.3.0
pandarallel==1.5.1
pandas==1.2.0
pandocfilters==1.4.3
parso==0.8.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plac==1.1.3
plotly==4.14.1
plotnine==0.7.1
pluggy==0.13.1
preshed==3.0.5
prometheus-client==0.9.0
prompt-toolkit==3.0.9
ptyprocess==0.7.0
py==1.10.0
py-readability-metrics==1.4.5
pycparser==2.20
Pygments==2.7.3
pyLDAvis==2.1.2
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.1
python-dateutil==2.8.1
pytz==2020.5
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
requests==2.25.1
retrying==1.3.3
scikit-learn==0.24.0
scipy==1.6.0
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
smart-open==4.1.0
sniffio==1.2.0
spacy==2.3.5
srsly==1.0.5
statsmodels==0.12.1
terminado==0.9.2
testpath==0.4.4
thinc==7.4.5
threadpoolctl==2.1.0
toml==0.10.2
torch==1.7.1
tornado==6.1
tqdm==4.55.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
wasabi==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
zipp==3.4.0
